requirements.txt
Django==2.2.10
djangorestframework==3.11.0
pip==20.0.2
pytz==2019.3
setuptools==45.1.0
sqlparse==0.3.0
wheel==0.34.2